Refactor JDK bug work around
Note that while this is breaking as now the new program will display the
warning message. At the same this is a requirement for relative location
configuration loading to work, which is a planned feature. The
'-relaunch' flag can still be supplied as the last argument to bypass
the message in case the user wants to.